UBUNTU: SAUCE: iptables/ip6tables: Ensure entries are fully copied to userspace
BugLink: http://bugs.launchpad.net/bugs/1691752

Changes to copy entries one by one result in not copying the
matches and the target to userpace, since these are not included
in the size of struct ipt_entry and struct ip6t_entry. This causes
a regression in deleting rules. Partially revert those changes to
fix this.

Fixes: f77bc5b23fb1 ("iptables: use match, target and data copy_to_user helpers")
Fixes: e47ddb2c4691 ("ip6tables: use match, target and data copy_to_user helpers")
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>