vmbus: don't acquire the mutex in vmbus_hvsock_device_unregister()
This is a tempoary workaround to avoid deadlock:

If we get the mutex in vmbus_hvsock_device_unregister() and call
vmbus_device_unregister() -> device_unregister() -> ... -> device_release()
-> vmbus_device_release(), we'll get a deadlock, because
vmbus_device_release() tries to get the same lock.

We need to add a proper sync mechanism for onoffer, onrescind and
vmbus_hvsock_device_unregister().

Signed-off-by: Dexuan Cui <decui@microsoft.com>
Cc: K. Y. Srinivasan <kys@microsoft.com>
Cc: Haiyang Zhang <haiyangz@microsoft.com>
Cc: Stephen Hemminger <sthemmin@microsoft.com>