netvsc: make sure and unregister datapath
Go back to switching datapath directly in the notifier callback.
No need for calling the NOTIFY_PEERS notifier; that is used to send
a gratitious ARP/ND packet; but that is not required with Hyper-V
because both VF and synthetic NIC have the same MAC address. Both
paths have same visible addresses.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>