netvsc: no need to allocate send/receive on numa node
BugLink: http://bugs.launchpad.net/bugs/1731994

The send and receive buffers are both per-device (not per-channel).
The associated NUMA node is a property of the CPU which is per-channel
therefore it makes no sense to force the receive/send buffer to be
allocated on a particular node (since it is a shared resource).

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 958333708f2877d3855e3bc31dad428e2f2c8096)
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>