UBUNTU: SAUCE: netvsc: defer initializing sub-channels
BugLink: http://bugs.launchpad.net/bugs/1731994

When a network device is hot added to a VM, the netvsc device would
hang waiting for host to open sub-channels. The problem is that the
probe routine was being called from processing of the primary channel
open request. The sub-channel open response would never arrive since
the primary channel offer message never completed.

The solution is to move the setup of sub-channels into a work queue
that gets triggered (if necessary) after primary channel is setup.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>