Drivers: hv: vmbus: do not mark HV_PCIE as perf_device
The pci-hyperv driver's channel callback hv_pci_onchannelcallback() is not
really a hot path, so we don't need to mark it as a perf_device, meaning
with this patch all HV_PCIE channels' target_cpu will be CPU0.

This patch can help to avoid a potential ABBA deadlock in the busy loop
of Hyper-V PCI driver's hv_compose_msi_msg().

Signed-off-by: Dexuan Cui <decui@microsoft.com>
Cc: stable@vger.kernel.org
Cc: Stephen Hemminger <sthemmin@microsoft.com>
Cc: K. Y. Srinivasan <kys@microsoft.com>