UBUNTU: [Packaging] update flavour-control.stub for signing changes
Based on the change for linux-azure in bionic:

When signing we now produce a linux-image-unsigned package, with
linux-signed producing the linux-image package with a signed
kernel. Along with this the base modules are split out into a
linux-modules package, and linux-image-extra is renamed to
linux-modules-extra. Update the control stub for these changes.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>