Tools: hv: kvp: Truncate the kvp_pool files if the files already exist
The files contain stale data after the VM image is copied from one host to
another and the VM starts to run on the new host. Of course, when the
daemon starts to run, it tries to update the files, but if the daemon
can't work (e.g. due to a kernel bug), the files may remain unmodified.

Another reason is that: when a VM image is copied from a new host to an
old host, the old host may not support some key-value pairs that are only
supported on new hosts.

Signed-off-by: Dexuan Cui <decui@microsoft.com>
Cc: K. Y. Srinivasan <kys@microsoft.com>
Cc: Haiyang Zhang <haiyangz@microsoft.com>
Cc: Stephen Hemminger <sthemmin@microsoft.com>
Cc: <Stable@vger.kernel.org>