Drivers: hv: vmbus: Remove the unsafe API vmbus_are_subchannels_present()
The test "list_empty(&primary->sc_list)" is a race condition.

The comment in the header file is incorrect: when a vmbus upper-layer
driver is unloaded, the host rescinds all the sub-channels related to
the driver, and the primary channel's primary->sc_list becomes empty.

We don't really need the API, since vmbus_onoffer() ->
vmbus_process_offer() -> channel->sc_creation_callback() can set up
the sub-channel(s) properly.

Now that we removed the last user of the API from storvsc, let's remove
the unsafe API.

Let's add a "Fixes" tag to indicate that this patch should only be applied
after that patch is used.

Fixes: d69a98c7b4bd ("scsi: storvsc: Fix a race in sub-channel creation that can cause panic")
Cc: Stephen Hemminger <sthemmin@microsoft.com>
Cc: K. Y. Srinivasan <kys@microsoft.com>
Cc: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: Dexuan Cui <decui@microsoft.com>