vmbus: fix subchannel removal
The changes to split ring allocation from open/close, broke
the cleanup of subchannels. This resulted in problems using
uio on network devices because the subchannel was left behind
when the network device was unbound.

The cause was in the disconnect logic which used list splice
to move the subchannel list into a local variable. This won't
work because the subchannel list is needed later during the
process of the rescind messages (relid2channel).

The fix is to just leave the subchannel list in place
which is what the original code did. The list is cleaned
up later when the host rescind is processed.

Fixes: ae6935ed7d42 ("vmbus: split ring buffer allocation from open")
Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>

The patch was posted on Dec-7 2018:
https://www.spinics.net/lists/linux-driver-devel/msg120802.html
but it has not been accepted in the upstream.

Now let me rebase it to v4.20: the line  "channel->num_sc = 0;"
must be removed due to
4d3c5c69191f ("Drivers: hv: vmbus: Remove the useless API vmbus_get_outgoing_channel()")

Signed-off-by: Dexuan Cui <decui@microsoft.com>