hv_balloon: Add the support of hibernation
When hibernation is enabled, we must ignore the balloon up/down and
hot-add requests from the host, if any.

In the thread dm_device.thread, convert the "struct completion" to
wait_queue_head_t, which supports the hibernation-aware API
wait_event_freezable_timeout(), so the kernel thread is automatically
frozen and thawed on hibernation, and then we don't need to destroy
and re-create the thread.

Signed-off-by: Dexuan Cui <decui@microsoft.com>