Drivers: hv: vmbus:  Hardcode MMIO resources in vmbus_walk_resources() when necessary
Currently Hyper-V has not provided valid MMIO resources for TDX VMs (without pavavisor).
Hardcode the resources as a temporary workaround.

This patch is tested on a local Hyper-V host for TDX VMs.

These hardcoded MMIO ranges are the same default values for a regular non-CVM VM
on a local Hyper-V host. On Azure, the default ranges are:
[c0000000, ffffffff] and [fe0000000, fffffffff]. These hardcoded MMIO ranges here
should also work when a TDX VM runs on Azure, though the patch uses a smaller 32-bit
MMIO range that's a subset of [c0000000, ffffffff].

The patch should not cause any regression for non-TDX VMs because the 'start' in
those VMs should be non-zero, so it should be safe to include this patch in an
official kernel.

Signed-off-by: Dexuan Cui <decui@microsoft.com>
(cherry picked from commit https://github.com/dcui/linux/commit/94d592eb7658add7da26709f43446b7d68e44b5a)
Signed-off-by: Dexuan Cui <decui@microsoft.com>